file-processor/requirements.txt
pydicom
pillow
